refactor: move get_contract_bytecode to a new bytecode module
