+++ common/src/ether/mod.rs
@@ -4,3 +4,4 @@ pub mod lexers;
 pub mod rpc;
 pub mod selectors;
 pub mod signatures;
+pub mod bytecode;
